requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
psycopg[binary]>=3.2,<4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qrcode==8.2
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.7
